minor error handling decoratorfixes, standardized and more descriptive exception docstrings
