+++ btrdb/endpoint.py
@@ -28,9 +28,8 @@
 from btrdb.grpcinterface import btrdb_pb2
 from btrdb.grpcinterface import btrdb_pb2_grpc
 from btrdb.point import RawPoint
-from btrdb.exceptions import BTrDBError
+from btrdb.exceptions import BTrDBError, error_handler, check_proto_stat
 from btrdb.utils.general import unpack_stream_descriptor
-from btrdb.errors import error_handler, check_proto_stat
 
 
 class Endpoint(object):
